feat: remove the edit function from the UX
